SPARKC-190: Save complex objects to Cassandra
This ticket introduces new MappedToGettableDataConverter class
that can convert any Scala object having a ColumnMapper to a
CassandraRow, TupleValue or UDTValue object, easy to write directly
to Cassandra. MappedToGettableDataConverter handles collections,
nested tuples and nested UDT values. It is also a standalone
component with a simple API and no dependencies on the Java Cassandra
driver classes, which makes it easily unit-testable without mocking.

The implementation of MTGDC has been based on the original
implementation taken from DefaultRowWriter and then extended with the
new features. DefaultRowWriter class delegates the conversion work
to MTGDC.